requirements.txt
streamlit
pandas
fpdf2
xlsxwriter
streamlit-mic-recorder
google-generativeai
python-dotenv
groq
openpyxl